Further additions to IkamoanaFields
Added a first attempt at loading the current data from .dym as an additional IkamoanaFields function to match taxis, diffusion etc.
Still need to cut the lon/lat/time coordinates down to match that has been specified by the user

Finally, I hardcoded and added some current file locations and data just for testing purposes.

I've also fixed something, which might not have been a bug:
the wrapper classes for calculating taxis only made a local copy of the feeding_habitat which was not saved as a class attribute. This means that calls to the diffusion function failed, because there is no habitat data.
I've changed this but @Ash12H take a look and make sure that this was not fait exprès!