Update ikamoana.py
Updated to using the new wrapper code from IkamoanaFields, rather than having instances of both that and FeedingHabitat as attributes to this class.

Also added functions to initiate a particle set from positions given by the user (still need some code to load starting distribution fields), and also to concatenate the JIT-mode kernels and run them on the particle set. The movement of the particles is clearly not correct yet, but the basic simulation structure is now all here.

Finally, there are some new xml file parameters read in and a wrapper class for adding constants to the fieldset, which will be important for mortality kernels later on.