Create IkaSim_Example.nc
A first, simple example of using the new IkaSim class to create an ikamoana simulation using the IkaSim_Example xml parameter file.

It is very simple, but shows the steps needed to create a typical Ikamoana simulation- loading the parameters that will point to seapodym forcings and parameters, creating the forcings, creating the particle sets, and then finally executing the advection loop across a short period of time.

More to add on this, but if parcels is installed this should now work (though something is up with our K forcings still!!)