Update behaviours.py
I spotted an error in the diffusion kernel. When rearranging from the previous ikamoana code, I realised I was double-dipping with the value of particle.dt in the lon and lat scalers-this is why values were enormous!

Still need to test these for accuracy, but the particle movement now appears to be at the right scale at least.