Particle starting distribution
Added a function to Ikamoana fields to load and dim-slice a tuna density .dym file specified by the user. At the moment we assume a naming convention based on the starting cohort age, but perhaps we will need to change this later.
 Similar to the function to load currents data, except this is not used as a 'forcing', rather a distribution from which to initiate particles from initialiseFishParticles().

A couple of bug fixes and superfluous code removed from IkaSim and a new argument added to the Ika Params xml file for the starting distribution