Landmask loading
First commit creating a landmask for loading into the parcels FieldSet for actual kernel landmasking, which requires that the dataarray has a temporal dimension.

Added an argument to the IkamoanaFields.landmask to specify returning an xarray.dataarry with a time dimension, which is arbitrarily set to 0 as absolute dates will be ignored for the landmask anyway (as it doesn't change through time).

How it is added to the forcing dictionary, and then loaded into parcels has been added to IkaSim.createFieldSet()