Initial field sampling
Needed to make a small addition to runKernels to allow the user to specify kernels that will be run with dt=0, _before_ the main execution loop. This allows initial field values to be sampled at time=0, as per the parcels tutorials on the subjet.
Also, small addition to configfilegenerator to return the final filepath object when write() is called.