Cell to start field changes
Small change to help cope with instances of user-requested start cells being landmask cells.

We now use the landmask to define the start_field grid, and check that the cell is not a land cell. If so, then a warning message is printed and those surrounding cells in the Moore neighbourhood that are also not land are chosen instead.