FAD Attraction kernels
Updated FAD attraction kernels with bugfixes and improved code.

There are still issues when particles leave the domain causing errors to be thrown inside the interaction kernels, but these should be easy to resolve.

More testing still required as I'm not 100$ displacements are occurring in the correct order within mutator functions, but committing for now