Unscaled feeding habitat and related sampler kernels
First major commit, which includes changes to feeding habitat, fieldsgenerator and ikamoanafields to create a duplicate feeding habitat field without scaling, representing the estimation of raw micronekton density in tonnes/km^2, which is then passed all the way with other fields to join the ocean fieldset.
Alongside this are some changes to generalise the loading of initial particle density to allow pointing to dym files with no time dimension, but an age_class dimension (as with the true SEAPODYM init file). In this case, the only time-step available will be used for the starting age class determined by the user.
Finally some additional sampling kernels were added to behaviours.py to record these fields over the trajectories.

TODO: Also pass the raw, integrated temperature field through to the ocean fieldset creation for recording by particles